+++ jest.config.js
@@ -3,7 +3,7 @@ const { pathsToModuleNameMapper } = require('ts-jest');
 const { compilerOptions } = require('./tsconfig');
 module.exports = {
   preset: 'ts-jest',
-  testEnvironment: 'node',
+  testEnvironment: 'jsdom',
   moduleNameMapper: {
     '\\.(css|less)$': '<rootDir>/src/util/jest_css_placeholder.tsx',
     ...pathsToModuleNameMapper(compilerOptions.paths, { prefix: "<rootDir>/src" })