+++ docs/changelog.md
@@ -3,6 +3,7 @@
 - first round of top-down-update code removal
 - fix focus traversal and child creation
 - fix focus after arrangement
+- fix keyboard focus issues
 
 # Version 3.6.0: Add omnibar 
 - Refactor to separate omnibar and search